refactor: refactor to components and pages and utils sections
